refactor: 관리자 지원현황 ssr 적용 (#71)
