+++ apps/web/src/app/(main)/apply-management/[tab]/[id]/page.tsx
@@ -1,67 +1,25 @@
-'use client';
-import { useParams, useRouter } from 'next/navigation';
-import type { documentEvaluation as DocumentEvaluationType } from '@web/types/document-evaluation';
-import { documentEvaluationDummyData } from '@web/constants/document-evaluation';
-import ApplicantDetail from '@web/app/(main)/apply-management/[tab]/[id]/_components/ApplicantDetail/ApplicantDetail';
-import * as styles from './page.css';
-import { EvaluationScoreCard } from '@web/app/(main)/apply-management/[tab]/[id]/_components/EvaluationScoreCard/EvaluationScoreCard';
-import { RelationCard } from '@web/app/(main)/apply-management/[tab]/[id]/_components/RelationCard/RelationCard';
-import { EvaluationCommentCard } from '@web/app/(main)/apply-management/[tab]/[id]/_components/EvaluationCommentCard/EvaluationCommentCard';
-import { DetailHeader } from '@web/app/(main)/apply-management/[tab]/[id]/_components/DetailHeader/DetailHeader';
-import { Flex } from '@repo/ui/Flex';
-import { useApplicationDetailQuery } from '@web/store/query/useApplicationDetailQuery';
+import React from 'react';
+import { getServerSideTokens } from '@web/api/serverSideTokens';
+import { ServerFetchBoundary } from '@web/store/query/ServerFetchBoundary';
+import { getApplicationDetailQueryOptions } from '@web/store/query/useApplicationDetailQuery';
+import ApplicationDetailClient from './ApplicationDetailClient';
 
-export default function Page() {
-  const { id, tab } = useParams() as { id: string; tab: string };
-  const applicationId = Number(id);
+interface PageProps {
+  params: { tab: string; id: string };
+}
 
-  const { data, isLoading, isError } = useApplicationDetailQuery(applicationId);
+export default async function Page({ params }: PageProps) {
+  const applicationId = Number(params.id);
+  const tokens = await getServerSideTokens();
 
-  console.log('관리자 지원서', data);
-  if (isError || !data) {
-    return <div>지원서 정보를 불러올 수 없습니다.</div>;
-  }
+  const fetchOptions = getApplicationDetailQueryOptions({
+    applicationId,
+    tokens,
+  });
 
   return (
-    <div className={styles.container}>
-      {/* 헤더 */}
-      <DetailHeader tab={tab} name={data.name} />
-      <Flex gap="2rem">
-        <ApplicantDetail application={data} />
-
-        <div className={styles.rightSection}>
-          <EvaluationScoreCard
-            evaluationType="document"
-            evaluation={data.evaluations
-              .filter((e) => e.criteria.type === 'DOCUMENT')
-              .map((e) => ({
-                evaluator: e.user.name,
-                status: 'complete', // 예시: score가 null이면 대기중, 있으면 완료로 처리
-                score: e.score ?? null,
-                color: e.user.profileColor,
-              }))}
-          />
-
-          <EvaluationScoreCard
-            evaluationType="interview"
-            evaluation={data.evaluations
-              .filter((e) => e.criteria.type === 'INTERVIEW')
-              .map((e) => ({
-                evaluator: e.user.name,
-                status: e.score != null ? 'complete' : 'pending',
-                score: e.score ?? null,
-                color: e.user.profileColor,
-              }))}
-          />
-          <RelationCard relations={data.acquaintances.map((a) => a.name)} />
-          <EvaluationCommentCard
-            comments={data.documentComments.map((c) => ({
-              evaluator: c.user.name,
-              comment: c.content,
-            }))}
-          />
-        </div>
-      </Flex>
-    </div>
+    <ServerFetchBoundary fetchOptions={fetchOptions}>
+      <ApplicationDetailClient tab={params.tab} applicationId={applicationId} />
+    </ServerFetchBoundary>
   );
 }
