+++ apps/web/src/app/(main)/apply-management/_components/Tab/DocumentTab/DocumentTab.tsx
@@ -101,6 +101,8 @@ export default function DocumentTab({
         smsSent: item.isSmsSent,
         mailSent: item.isMailSent,
         evaluators: item.documentEvaluators.map((e) => ({
+          userId: e.userId,
+          profileImageUrl: e.profileImageUrl,
           name: e.name,
           profileColor: e.profileColor,
         })),
