+++ apps/web/src/app/(main)/apply-management/_components/Tab/FinalTab/FinalTab.tsx
@@ -94,6 +94,8 @@ export default function FinalTab({
         smsSent: item.isSmsSent,
         mailSent: item.isMailSent,
         evaluators: item.documentEvaluators.map((e) => ({
+          userId: e.userId,
+          profileImageUrl: e.profileImageUrl,
           name: e.name,
           profileColor: e.profileColor,
         })),
