+++ apps/web/src/app/(main)/apply-management/_components/Tab/RejectedTab/RejectedTab.tsx
@@ -94,6 +94,8 @@ export default function RejectedTab({
         smsSent: item.isSmsSent,
         mailSent: item.isMailSent,
         evaluators: item.documentEvaluators.map((e) => ({
+          userId: e.userId,
+          profileImageUrl: e.profileImageUrl,
           name: e.name,
           profileColor: e.profileColor,
         })),
