+++ apps/web/src/app/(main)/apply-management/add/_components/BasicInfoForm/BasicInfoForm.tsx
@@ -16,7 +16,6 @@ import Image from 'next/image';
 import * as styles from './BasicInfoForm.css';
 import { Flex } from '@repo/ui/Flex';
 import * as styles1 from '../../../../application-list/setting/(preview)/_components/AdditionalInfoPreview/AdditionalInfoPreview.css';
-import * as s from '../../../../application-list/setting/(preview)/_components/BasicInfoPreview/BasicInfoPreview.css';
 import { useFormFieldStatus } from '@web/app/apply/[organization]/[slug]/_context/FormFieldStatusContext';
 import { focusableWrapper } from '@web/app/apply/[organization]/[slug]/_components/FormNavigator/FormNavigator.css';
