+++ apps/web/src/app/(main)/apply-management/add/_components/InterviewScheduleForm/InterviewScheduleForm.tsx
@@ -29,7 +29,12 @@ export function InterviewScheduleForm({
   const scheduleStatus = useFormFieldStatus('interview-schedule');
 
   return (
-    <div id="interview-schedule" tabIndex={-1} className={focusableWrapper}>
+    <div
+      id="interview-schedule"
+      tabIndex={-1}
+      className={focusableWrapper}
+      style={{ width: '100%' }}
+    >
       <Flex gap="0.4rem" direction="column">
         <Flex gap="0.4rem">
           <Text variant="md1_text_semibold" color="grayscale70">
