+++ apps/web/src/store/query/useLatestDistribution.ts
@@ -6,6 +6,7 @@ import {
 import { GET } from '@web/api/fetch';
 import { HTTPError } from 'ky';
 import { queryKeys } from '../constants';
+import { Tokens } from '@web/api/types';
 
 export interface AssignmentItem {
   positionName: string;
@@ -25,14 +26,20 @@ const LATEST_DIST_GC_TIME = 1000 * 60 * 5;
 
 export type GetLatestDistributionParams = {
   recruitmentId: number;
+  tokens?: Tokens;
 };
 
 /**
  * 최신 분배 정보를 가져올 쿼리 옵션
  * 404 에러는 null 로 처리
  */
+/**
+ * 최신 분배 정보를 가져올 쿼리 옵션
+ * 404 에러는 null 로 처리
+ */
 export function getLatestDistributionQueryOptions({
   recruitmentId,
+  tokens,
 }: GetLatestDistributionParams): UseSuspenseQueryOptions<
   LatestDistribution | null,
   Error
@@ -42,14 +49,18 @@ export function getLatestDistributionQueryOptions({
     queryFn: async () => {
       try {
         const res = await GET<LatestDistribution>(
-          `api/v1/admin/applications/distribute-evaluators/latest/${recruitmentId}`
+          `api/v1/admin/applications/distribute-evaluators/latest/${recruitmentId}`,
+          undefined,
+          tokens
         );
         return res.result;
-      } catch (err) {
-        if (err instanceof HTTPError && err.response.status === 404) {
+      } catch (err: any) {
+        // response.status가 404면 null 리턴
+        if (err?.response?.status === 404) {
           return null;
         }
-        throw err as Error;
+        // 그 외 에러는 다시 던져서 React Query가 처리하게
+        throw null;
       }
     },
     staleTime: LATEST_DIST_STALE_TIME,
@@ -59,6 +70,8 @@ export function getLatestDistributionQueryOptions({
   });
 }
 
-export function useLatestDistributionQuery(recruitmentId: number) {
-  return useSuspenseQuery(getLatestDistributionQueryOptions({ recruitmentId }));
+export function useLatestDistributionQuery(
+  params: GetLatestDistributionParams
+) {
+  return useSuspenseQuery(getLatestDistributionQueryOptions(params));
 }
