+++ apps/web/src/store/query/useRecruitmentPositionsQuery.ts
@@ -5,6 +5,7 @@ import {
 } from '@tanstack/react-query';
 import { GET } from '@web/api/fetch';
 import { queryKeys } from '../constants';
+import { Tokens } from '@web/api/types';
 
 /**
  * GET /api/v1/positions/recruitment/{recruitmentId}
@@ -19,25 +20,34 @@ export interface Position {
 const STALE_TIME = 1000 * 60 * 2;       // 1분
 const GC_TIME    = 1000 * 60 * 3;  // 1시간
 
-export function getRecruitmentPositionsQueryOptions(
-  recruitmentId: number
-): UseSuspenseQueryOptions<Position[], Error> {
+export type RecruitmentPositionsParams = {
+  recruitmentId: number;
+  tokens?: Tokens;
+};
+
+export function getRecruitmentPositionsQueryOptions({
+  recruitmentId,
+  tokens,
+}: RecruitmentPositionsParams): UseSuspenseQueryOptions<Position[], Error> {
   return queryOptions<Position[]>({
     queryKey: queryKeys.positions.byRecruitment(recruitmentId),
     queryFn: () =>
       GET<Position[]>(
-        `api/v1/positions/recruitment/${recruitmentId}`
+        `api/v1/positions/recruitment/${recruitmentId}`,
+        undefined,
+        tokens           // ← 여기로 토큰 전달
       ).then((res) => res.result),
     staleTime: STALE_TIME,
-    gcTime: GC_TIME,
-    enabled: recruitmentId > 0,
+    gcTime:   GC_TIME,
+    enabled:  recruitmentId > 0,
   });
 }
 
 export function useRecruitmentPositionsQuery(
-  recruitmentId: number
+  recruitmentId: number,
+  tokens?: Tokens
 ) {
   return useSuspenseQuery(
-    getRecruitmentPositionsQueryOptions(recruitmentId)
+    getRecruitmentPositionsQueryOptions({ recruitmentId, tokens })
   );
 }
